+++ src/odin/maneuversbuilder.cc
@@ -1360,6 +1360,10 @@ void ManeuversBuilder::FinalizeManeuver(Maneuver& maneuver, int node_index) {
     }
   }
 
+  if (curr_edge->IsPedestrianCrossingUse() && prev_edge && prev_edge->IsFootwayUse()) {
+    maneuver.set_pedestrian_crossing(true);
+  }
+
   // Set the maneuver type
   SetManeuverType(maneuver);
 }
@@ -2261,6 +2265,14 @@ bool ManeuversBuilder::IsPedestrianFork(int node_index,
         node->GetStraightestTraversableIntersectingEdgeTurnDegree(prev_edge->end_heading(),
                                                                   prev_edge->travel_mode(),
                                                                   &xedge_use);
+
+    // This is needed to ensure we don't consider footways and crosswalks as
+    // different uses for determining pedestrian forks
+    bool is_current_and_intersecting_edge_of_similar_use =
+        (curr_edge->use() == xedge_use ||
+         (curr_edge->IsFootwayUse() &&
+          (xedge_use == TripLeg_Use_kPedestrianCrossingUse || xedge_use == TripLeg_Use_kFootwayUse)));
+
     // if there is a similar traversable intersecting edge
     //    or there is a relative straight traversable intersecting edge
     //    and the current edge use has to be the same as the intersecting edge use
@@ -2269,7 +2281,7 @@ bool ManeuversBuilder::IsPedestrianFork(int node_index,
     if (((((xedge_counts.left_similar_traversable_outbound > 0) ||
            (xedge_counts.right_similar_traversable_outbound > 0)) ||
           is_relative_straight(straightest_traversable_xedge_turn_degree)) &&
-         (curr_edge->use() == xedge_use)) ||
+         is_current_and_intersecting_edge_of_similar_use) ||
         (prev_edge->roundabout() && !curr_edge->roundabout())) {
       return true;
     }
@@ -2788,10 +2800,11 @@ void ManeuversBuilder::ProcessRoundabouts(std::list<Maneuver>& maneuvers) {
           // suppressed
           curr_man->set_roundabout_exit_begin_heading(next_man->begin_heading());
 
-          // Store the next maneuver's turn degree. This is where
+          // Store the next maneuver's turn degree and exit shape index. This is where
           // we store the turn angles when exit roundabout maneuver is
           // suppressed
           curr_man->set_roundabout_exit_turn_degree(next_man->turn_degree());
+          curr_man->set_roundabout_exit_shape_index(curr_man->end_shape_index());
 
           // Set the traversable_outbound_intersecting_edge booleans
           curr_man->set_has_left_traversable_outbound_intersecting_edge(