+++ src/thor/triplegbuilder.cc
@@ -328,6 +328,9 @@ void SetShapeAttributes(const AttributesController& controller,
     }
     distance_total_pct = next_total;
     double time = distance / cut_itr->speed; // seconds
+    if (std::isnan(time)) {
+      time = 0.;
+    }
 
     // Set shape attributes time per shape point if requested
     if (controller.attributes.at(kShapeAttributesTime)) {
@@ -344,7 +347,11 @@ void SetShapeAttributes(const AttributesController& controller,
     // Set shape attributes speed per shape point if requested
     if (controller.attributes.at(kShapeAttributesSpeed)) {
       // convert speed to decimeters per sec and then round to an integer
-      leg.mutable_shape_attributes()->add_speed((distance * kDecimeterPerMeter / time) + 0.5);
+      double speed = (distance * kDecimeterPerMeter / time) + 0.5;
+      if (std::isnan(speed) || time == 0.) { // avoid NaN
+        speed = 0.;
+      }
+      leg.mutable_shape_attributes()->add_speed(speed);
     }
 
     // Set the maxspeed if requested
@@ -860,6 +867,10 @@ TripLeg_Edge* AddTripEdge(const AttributesController& controller,
 
   if (controller.attributes.at(kEdgeSacScale)) {
     trip_edge->set_sac_scale(GetTripLegSacScale(directededge->sac_scale()));
+  }
+
+  if (controller.attributes.at(kEdgeShoulder)) {
+    trip_edge->set_shoulder(directededge->shoulder());
   }
 
   if (controller.attributes.at(kEdgeSidewalk)) {