+++ valhalla/odin/enhancedtrippath.h
@@ -379,6 +379,7 @@ public:
   bool IsPathUse() const;
   bool IsPedestrianUse() const;
   bool IsBridlewayUse() const;
+  bool IsPedestrianCrossingUse() const;
   bool IsRestAreaUse() const;
   bool IsServiceAreaUse() const;
   bool IsOtherUse() const;